+++ app/Controllers/Http/ItemsController.ts
@@ -1,7 +1,7 @@
-import {HttpContextContract} from '@ioc:Adonis/Core/HttpContext'
+import { HttpContextContract } from "@ioc:Adonis/Core/HttpContext";
 import Item from "App/Models/Item";
 import ItemValidator from "App/Validators/item/ItemValidator";
-import Logger from "@ioc:Adonis/Core/Logger";
+import type { Item as IItem } from "../../../types";
 
 export default class ItemsController {
   /**
@@ -10,15 +10,11 @@ export default class ItemsController {
    * @param request
    * @return Promise<Item>
    */
-  public async store({auth, request}: HttpContextContract) {
+  public async store({ auth, request }: HttpContextContract) {
     await auth.use("api").authenticate();
-    try {
-      const payload = await request.validate(ItemValidator);
-      // @ts-ignore
-      return await Item.create(payload);
-    } catch (e) {
-      Logger.warn(e);
-    }
+    const payload = await request.validate(ItemValidator);
+    // @ts-ignore
+    return await Item.create(payload);
   }
 
   /**
@@ -27,27 +23,51 @@ export default class ItemsController {
    * @param auth
    * @return Promise<Item[]>
    */
-  public async index({ params, auth }: HttpContextContract){
-    await auth.use("api").authenticate()
-    return Item.query().select("*").where("category", params.category)
+  public async index({ params, auth }: HttpContextContract) {
+    await auth.use("api").authenticate();
+    return Item.query().select("*").where("category", params.category);
   }
 
-  public async update({ params, auth }: HttpContextContract) {
+  public async update({ params, auth, request }: HttpContextContract) {
     await auth.use("api").authenticate();
-    const item = (await Item.find(params.id))!
-    item.state = !item.state
-    return item.save()
+    const item = (await Item.find(params.id))!;
+
+    const body: IItem = request.body();
+    const keys = Object.keys(body);
+    const source = ["name", "description", "category", "start", "end"];
+
+    if (keys.map((item) => source.includes(item)).includes(true)) {
+      keys.forEach((key) => {
+        if (source.includes(key)) {
+          item[key] = body[key];
+        }
+      });
+      return item.save();
+    }
+
+    item.state = !item.state;
+    return item.save();
   }
 
   public async delete({ auth, params }) {
     await auth.use("api").authenticate();
-    const item = (await Item.find(params.id))!
-    return item.delete()
+    const item = await Item.find(params.id);
+    if (item === null) {
+      return {
+        error: {
+          message: `Item with id: ${params.id} is not found.`,
+        },
+      };
+    }
+    await item.delete();
+    return {
+      message: `Item with id: ${params.id} has been deleted.`,
+    };
   }
 
   public async get({ auth }: HttpContextContract) {
     await auth.use("api").authenticate();
-    const {id} = auth.user!;
+    const { id } = auth.user!;
     return Item.query().select("*").where("owner", id);
   }
 }
